Large refactor of map and sidenav components
This refactor has several goals some of which are still
works-in-progress:
- Separate the map and sidenav into logically-separate components
- Make components primarily focussed on presentation and move other
logic into services.
- Share state between components by injecting services into
components as needed.
- Use immutable data structures/observables when modifying state
based on a user-interaction event.
