+++ frontend/src/app/app.component.html
@@ -1,3 +1,6 @@
 <psu-campusmap-toolbar></psu-campusmap-toolbar>
 <psu-campusmap-search></psu-campusmap-search>
-<psu-campusmap-map></psu-campusmap-map>
+<md-sidenav-container>
+  <sidenav></sidenav>
+  <psu-campusmap-map></psu-campusmap-map>
+</md-sidenav-container>
