+++ frontend/src/styles.scss
@@ -24,3 +24,9 @@ md-toolbar a {
     margin: 0;
     padding: 16px;
 }
+
+md-sidenav-container {
+    top: $toolbar-height;
+    width: 100%;
+    height: 100%;
+}
